research_sandbox/requirements.txt
numpy
scipy
scikit-learn
matplotlib
pyyaml
pytest
numba
